reorganizando os componentes e atualizando o readme
