+++ README.md
@@ -1,2 +1,4 @@
-# casa-do-codigo
+# Casa do código
 Projeto de um CRUD para a casa do código, aplicação desenvolvida durante o curso da Alura em React
+
+![casa-do-codigo-layout](/src/assets/img/casa-do-codigo.png)
