+++ src/components/TableHead.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const TableHead = () => {
+    return (
+        <thead>
+            <tr>
+                <th>Autores</th>
+                <th>Livros</th>
+                <th>Preços</th>
+                <th>Remover</th>
+            </tr>
+        </thead>
+    );
+}
+
+export default TableHead;
